feat(refactor): refactored project structure to make it more scalable and testable
