+++ frontend/index.html
@@ -17,7 +17,7 @@
   <title>car-shop</title>
 </head>
 <body>
-  <car-shop></car-shop>
+  <app-root></app-root>
   <script type="module" src="./index.js"></script>
   <script>
     const process = {
