+++ frontend/index.js
@@ -1,6 +1,22 @@
-import { CarShop, CarsList, CarCard, CarDetails } from './packages/car-shop';
+import {
+  html,
+  LitElement,
+  registerDefaultIconsets,
+} from 'ing-web';
 
-customElements.define('car-shop', CarShop);
-customElements.define('cars-list', CarsList);
-customElements.define('car-card', CarCard);
-customElements.define('car-details', CarDetails);
+import './pages/car-shop';
+
+registerDefaultIconsets();
+
+export class App extends LitElement {
+  constructor() {
+    super()
+  }
+  render() {
+    return html`
+      <car-shop></car-shop>
+    `
+  }
+}
+
+customElements.define('app-root', App);
