--- frontend/packages/car-shop/__element-definitions/car-card.js
@@ -1,3 +1,0 @@
-import { CarCard } from '../src/CarCard.js';
-
-window.customElements.define('car-card', CarCard);
