--- frontend/packages/car-shop/__element-definitions/car-details.js
@@ -1,3 +1,0 @@
-import { CarDetails } from '../src/CarDetails.js';
-
-window.customElements.define('car-details', CarDetails);
