--- frontend/packages/car-shop/__element-definitions/car-shop.js
@@ -1,3 +1,0 @@
-import { CarShop } from '../src/CarShop.js';
-
-window.customElements.define('car-shop', CarShop);
