--- frontend/packages/car-shop/index.js
@@ -1,4 +1,0 @@
-export { CarShop } from './src/CarShop.js';
-export { CarsList } from './src/CarsList.js';
-export { CarCard } from './src/CarCard.js';
-export { CarDetails } from './src/CarDetails.js';
