+++ frontend/packages/components/car-list/__element-definitions/cars-list.js
@@ -1,3 +1,3 @@
-import { CarsList } from '../src/CarsList.js';
+import { CarsList } from '../CarsList.js';
 
 window.customElements.define('cars-list', CarsList);
