+++ frontend/packages/components/car-list/car-card/__element-definitions/car-card.js
@@ -1,0 +1,3 @@
+import { CarCard } from '../../../../pages/car-shop/src/CarCard.js';
+
+window.customElements.define('car-card', CarCard);
