+++ frontend/packages/components/index.js
@@ -1,0 +1,4 @@
+export { ShoppingBasket } from './shopping-basket/ShoppingBasket';
+export { CarsList } from './car-list/CarsList';
+export { CarDetails } from './car-details/CarDetails';
+export { DialogWindow } from './dialog/dialog';
