+++ frontend/packages/dialogs/index.js
@@ -1,0 +1,1 @@
+export { CarDetailsDialog } from './car-details/CarDetailsDialog'
