+++ frontend/packages/store/modules/car-shop/index.js
@@ -1,0 +1,2 @@
+import carReducer from './reducer';
+export default carReducer;
