+++ frontend/packages/store/modules/index.js
@@ -1,0 +1,7 @@
+import carReducer from './car-shop';
+
+const appReducer = {
+  carReducer
+}
+
+export default appReducer;
