+++ frontend/packages/views/index.js
@@ -1,0 +1,1 @@
+export { CarShopView } from './car-shop/CarShopView'
