+++ frontend/pages/car-shop/index.js
@@ -1,0 +1,4 @@
+export { CarShopContainer } from './src/CarShopContainer.js';
+
+
+
